requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.7.1
psycopg[binary]>=3.3.0
python-dotenv==1.1.0
